finished project / finished adding tables for items and edit, delete, create implementation

+++ inventory/resources/views/categories/index.blade.php
@@ -5,10 +5,11 @@
       <div class="mask d-flex align-items-center h-100">
         <div class="container">
           <div class="row justify-content-center">
-            <div class="col-12">
+            <div class="col-7">
               <div class="card">
                 <div class="card-body p-0">
                   <div class="table-responsive table-scroll" data-mdb-perfect-scrollbar="true" style="position: relative; height: 700px">
+                    <h1 class="pull-right"><a href='/item/create' class='btn btn-info' role='button'>Create new Product</a></h1>
                     <h1 class="pull-right"><a href='/categories/create' class='btn btn-info' role='button'>Create new row</a></h1>
                     <table class="table table-striped mb-0">
                       <thead style="background-color: #ffffff;">
@@ -26,7 +27,8 @@
                           <td>{{ $category->category }}</td>
                           <td><a href="{{ route('categories.edit', $category->id) }}" class="btn btn-primary btn-sm">Edit</a></td>
                           <td>
-                            {!! Form::open(['route' => ['categories.destroy', $category->id], 'method' => 'DELETE']) !!}
+                            {!! Form::open(['route' => ['categories.destroy', $category->id], 'method' => 'DELETE',  'onsubmit' => 'return confirm("Delete this category? Y/N")']) !!}
+
                             {{ Form::submit('Delete', ['class' => 'btn btn-danger btn-sm']) }}
                             {!! Form::close() !!}
                           </td>
